requirements.txt
cohere
oci
oracledb
pillow
numpy
python-dotenv
gradio
